Other: Extract common utils from commands and move commands to own module.
